+++ backend/user/serializers.py
@@ -1,0 +1,22 @@
+from rest_framework import serializers
+from django.contrib.auth.models import User
+
+class UserRegisterSerializer(serializers.ModelSerializer):
+    password = serializers.CharField(write_only=True)
+    re_password = serializers.CharField(write_only=True)
+
+    class Meta:
+        model = User
+        fields = ['username', 'email', 'password', 're_password']
+
+    def validate(self, attrs):
+        if attrs['password'] != attrs['re_password']:
+            raise serializers.ValidationError({"password": "Passwords do not match."})
+        return attrs
+
+    def create(self, validated_data):
+        validated_data.pop('re_password')  # Remove re_password from validated data
+        user = User(**validated_data) 
+        user.set_password(validated_data['password'])  # Hash the password
+        user.save()
+        return user
